+++ tests/trainer/flags/test_fast_dev_run.py
@@ -1,5 +1,4 @@
 import os
-from unittest import mock
 from unittest.mock import Mock
 
 import pytest
@@ -30,7 +29,6 @@ def test_skip_on_fast_dev_run_tuner(tmpdir, tuner_alg):
 
 
 @pytest.mark.parametrize("fast_dev_run", [1, 4])
-@mock.patch.dict(os.environ, {"PL_DEV_DEBUG": "1"})
 def test_callbacks_and_logger_not_called_with_fastdevrun(tmpdir, fast_dev_run):
     """
     Test that ModelCheckpoint, EarlyStopping and Logger are turned off with fast_dev_run
@@ -70,6 +68,7 @@ def test_callbacks_and_logger_not_called_with_fastdevrun(tmpdir, fast_dev_run):
     checkpoint_callback = ModelCheckpoint()
     checkpoint_callback.save_checkpoint = Mock()
     early_stopping_callback = EarlyStopping()
+    early_stopping_callback._evaluate_stopping_criteria = Mock()
     trainer_config = dict(
         default_root_dir=tmpdir,
         fast_dev_run=fast_dev_run,
@@ -104,7 +103,7 @@ def test_callbacks_and_logger_not_called_with_fastdevrun(tmpdir, fast_dev_run):
 
         # early stopping should not have been called with fast_dev_run
         assert trainer.early_stopping_callback == early_stopping_callback
-        assert len(trainer.dev_debugger.early_stopping_history) == 0
+        early_stopping_callback._evaluate_stopping_criteria.assert_not_called()
 
     train_val_step_model = FastDevRunModel()
     trainer = Trainer(**trainer_config)